pyproject.toml
[tool.poetry]
name = "discordaudioreactor"
version = "0.1.0"
description = "**DiscordAudioReactor** is a Discord bot designed to listen to conversations in text channels and play specific audio files when certain keywords are detected. This bot adds a fun and interactive element to your Discord server by providing audio reactions based on user-defined keywords."
authors = ["Dawid <MDDawid0323@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pynacl = "^1.5.0"
discord-py = "^2.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
